+++ Public/Test-EnvironmentVariable.ps1
@@ -1,0 +1,33 @@
+function Test-EnvironmentVariable {
+  [CmdletBinding()]
+  Param (
+    [Parameter (
+      Position    = 1,
+      Mandatory   = $true,
+      HelpMessage = "Name of the environment variable"
+    )]
+    [ValidateNotNullOrEmpty()]
+    [String]
+    $Variable,
+    [Parameter (
+      Position    = 2,
+      Mandatory   = $false,
+      HelpMessage = "Scope of the environment variable"
+    )]
+    [ValidateSet ("Machine", "Process", "User")]
+    [String]
+    $Scope = "Machine"
+  )
+  Begin {
+    # Get global preference vrariables
+    Get-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
+  }
+  Process {
+    # Check if variable is defined
+    if (Get-EnvironmentVariable -Variable $Variable -Scope $Scope) {
+      return $true
+    } else {
+      return $false
+    }
+  }
+}