+++ CHANGELOG.md
@@ -4,6 +4,37 @@ All notable changes to the [PSTK](https://github.com/Akaizoku/PSTK) project will
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+The following functions have been added:
+-   Expand-CompressedFile
+-   Find-Key
+-   Get-CallerPreference
+-   Get-EnvironmentVariable
+-   Get-HTTPStatus
+-   Get-KeyValue
+-   Get-Path
+-   Remove-EnvironmentVariable
+-   Resolve-Array
+-   Resolve-URI
+-   Select-XMLNode
+-   Set-EnvironmentVariable
+-   Test-EnvironmentVariable
+-   Test-Service
+-   Update-File
+-   Write-ErrorLog
+
+### Changed
+The following functions have been updated:
+-   Compare-Hashtable
+-   Compare-Properties
+-   Complete-RelativePath
+-   Convert-FileEncoding
+-   Get-Properties
+-   Test-SQLConnection
+-   Write-Log
+
 ## [1.1.0](https://github.com/Akaizoku/PSTK/releases/tag/1.1.0) - 2018-10-15
 
 ### Added